requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary
pgvector==0.2.4
sentence-transformers==2.2.2
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.0 